add Data group and subcommands
the data group only has one subcommand:
ilab data generate

Moved the file structure of generation so it makes more sense and lives within the data group.
