+++ src/instructlab/data/data.py
@@ -1,0 +1,26 @@
+# SPDX-License-Identifier: Apache-2.0
+
+# Third Party
+from click_didyoumean import DYMGroup
+import click
+
+# First Party
+from instructlab import config
+
+# Local
+from .generate import generate
+
+
+@click.group(cls=DYMGroup)
+@click.version_option(package_name="instructlab")
+@click.pass_context
+def data(ctx):
+    """Command Group for Interacting with the Data generated by InstructLab.
+
+    If this is your first time running ilab, it's best to start with `ilab init` to create the environment.
+    """
+    ctx.obj = ctx.parent.obj
+    ctx.default_map = ctx.parent.default_map
+
+
+data.add_command(generate)
